[BACKEND] Refactored ToDoService to follow SOLID principles
