+++ backend/src/main/java/com/springboot/todoapp_backend/service/ToDoService.java
@@ -1,6 +1,5 @@
 package com.springboot.todoapp_backend.service;
 
-import com.springboot.todoapp_backend.Utilities.Constants;
 import com.springboot.todoapp_backend.dtos.NewToDoDTO;
 import com.springboot.todoapp_backend.dtos.UpdateToDoDTO;
 import com.springboot.todoapp_backend.model.ToDo;
@@ -10,29 +9,32 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.Duration;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Service
 public class ToDoService {
     private static final Logger logger = LoggerFactory.getLogger(ToDoService.class);
 
     private final List<ToDo> toDoList;
+    private final TodoValidationService validationService;
+    private final TodoSearchService searchService;
+    private final TodoStatsService statsService;
 
-    public ToDoService(){
+    public ToDoService(
+            TodoValidationService validationService,
+            TodoSearchService searchService,
+            TodoStatsService statsService
+    ) {
         this.toDoList = new ArrayList<>();
+        this.validationService = validationService;
+        this.searchService = searchService;
+        this.statsService = statsService;
     }
 
     public Optional<ToDo> getItem(String id) {
-        Optional<ToDo> optional = Optional.empty();
-        for (ToDo item : toDoList){
-            if(item.getId().equals(id)){
-                optional = Optional.of(item);
-                return optional;
-            }
-        }
-        return optional;
+        return toDoList.stream()
+                .filter(item -> item.getId().equals(id))
+                .findFirst();
     }
 
     public List<ToDo> getFilteredList(
@@ -43,20 +45,7 @@ public class ToDoService {
             String sortBy,
             String order
     ) {
-        int adjustedPage = page > 0 ? page - 1 : 0;
-        return toDoList.stream()
-                // Text filter
-                .filter(todo -> text == null || todo.getText().toLowerCase().contains(text.toLowerCase()))
-                // Priority filter
-                .filter(todo -> priority == null || todo.getPriority() == priority)
-                // Status filter
-                .filter(todo -> isDone == null || todo.isDone() == isDone)
-                // Sort by priority and/or dueDate
-                .sorted(getComparator(sortBy, order))
-                // Pagination
-                .skip((long) adjustedPage * Constants.PAGE_SIZE)
-                .limit(Constants.PAGE_SIZE)
-                .collect(Collectors.toList());
+        return searchService.getFilteredList(toDoList, text, priority, isDone, page, sortBy, order);
     }
 
     public Integer getTotalItems(
@@ -65,41 +54,16 @@ public class ToDoService {
             Boolean isDone,
             String sortBy
     ) {
-        List<ToDo> filteredTodos = toDoList.stream()
-                .filter(todo -> text == null || todo.getText().toLowerCase().contains(text.toLowerCase()))
-                .filter(todo -> priority == null || todo.getPriority() == priority)
-                .filter(todo -> isDone == null || todo.isDone() == isDone)
-                .sorted(getComparator(sortBy, "desc"))
-                .toList();
-
-        return filteredTodos.size();
+        return searchService.getTotalItems(toDoList, text, priority, isDone);
     }
 
-    private Comparator<ToDo> getComparator(String sortBy, String order) {
-        Comparator<ToDo> comparator;
-
-        if (sortBy != null && sortBy.equalsIgnoreCase("duedate")) {
-            comparator = Comparator.comparing(ToDo::getDueDate, Comparator.nullsLast(Comparator.naturalOrder()));
-        } else {
-            comparator = Comparator.comparing(ToDo::getPriority);
-        }
-
-        if ("asc".equalsIgnoreCase(order)) {
-            comparator = comparator.reversed();
-        }
-        //If sortBy is null || priority || invalid
-        return comparator;
-    }
-
-    public ToDo addItem(NewToDoDTO newToDo){
+    public ToDo addItem(NewToDoDTO newToDo) {
+        validationService.validateNewTodo(newToDo);
+        
         LocalDate dueDate = Optional.ofNullable(newToDo.getDueDate())
                 .filter(d -> !d.isEmpty())
                 .map(LocalDate::parse)
                 .orElse(null);
-
-        if (dueDate != null && dueDate.isBefore(LocalDate.now())) {
-            throw new IllegalArgumentException("The due date must be today or a future date.");
-        }
 
         ToDo newItem = ToDo.builder()
                 .text(newToDo.getText())
@@ -108,22 +72,15 @@ public class ToDoService {
                 .build();
 
         toDoList.add(newItem);
-
         logger.info("New item created with ID: {}", newItem.getId());
         return newItem;
     }
 
-
-
     public Optional<ToDo> updateItem(String id, UpdateToDoDTO request) {
+        validationService.validateUpdateTodo(request);
         Optional<ToDo> existingItem = getItem(id);
 
-        if (request.getText() == null && request.getPriority() == null) {
-            throw new IllegalArgumentException("No valid fields provided for update");
-        }
-
         existingItem.ifPresent(todo -> {
-
             if (Objects.nonNull(request.getText())) {
                 todo.setText(request.getText());
             }
@@ -133,15 +90,11 @@ public class ToDoService {
             if (request.getDueDate() == null || request.getDueDate().isEmpty()) {
                 todo.setDueDate(null);
             } else {
-                LocalDate newDueDate = LocalDate.parse(request.getDueDate());
-                if (newDueDate.isBefore(LocalDate.now())) {
-                    throw new IllegalArgumentException("The due date must be today or a future date.");
-                } else {
-                    todo.setDueDate(newDueDate);
-                }
+                todo.setDueDate(LocalDate.parse(request.getDueDate()));
             }
             logger.info("Item updated with ID: {}", id);
         });
+        
         return existingItem;
     }
 
@@ -181,75 +134,7 @@ public class ToDoService {
         return existingItem;
     }
 
-    public Map<String, String> getCompletionStats(){
-        long totalMillis = 0;
-        long totalHighPriorityMillis = 0;
-        long totalMediumPriorityMillis = 0;
-        long totalLowPriorityMillis = 0;
-
-        int totalTasks = 0;
-        int highPriorityTasks = 0;
-        int mediumPriorityTasks = 0;
-        int lowPriorityTasks = 0;
-
-        for (ToDo todo : toDoList) {
-            if (todo.isDone() && todo.getDoneDate() != null && todo.getCreationDate() != null) {
-                Duration duration = Duration.between(todo.getCreationDate(), todo.getDoneDate());
-
-                long millisTaken = duration.toMillis();
-
-                totalMillis += millisTaken;
-                totalTasks++;
-
-                switch (todo.getPriority()) {
-                    case HIGH:
-                        totalHighPriorityMillis += millisTaken;
-                        highPriorityTasks++;
-                        break;
-                    case MEDIUM:
-                        totalMediumPriorityMillis += millisTaken;
-                        mediumPriorityTasks++;
-                        break;
-                    case LOW:
-                        totalLowPriorityMillis += millisTaken;
-                        lowPriorityTasks++;
-                        break;
-                }
-            }
-        }
-
-        long averageMillis = totalTasks > 0 ? totalMillis / totalTasks : 0;
-        long averageHighPriorityMillis = highPriorityTasks > 0 ? totalHighPriorityMillis / highPriorityTasks : 0;
-        long averageMediumPriorityMillis = mediumPriorityTasks > 0 ? totalMediumPriorityMillis / mediumPriorityTasks : 0;
-        long averageLowPriorityMillis = lowPriorityTasks > 0 ? totalLowPriorityMillis / lowPriorityTasks : 0;
-
-        Map<String, String> result = new HashMap<>();
-        result.put("averageTime", formatMillisToStandardTime(averageMillis));
-        result.put("averageTimeHighPriority", formatMillisToStandardTime(averageHighPriorityMillis));
-        result.put("averageTimeMediumPriority", formatMillisToStandardTime(averageMediumPriorityMillis));
-        result.put("averageTimeLowPriority", formatMillisToStandardTime(averageLowPriorityMillis));
-
-
-        return result;
-
+    public Map<String, String> getCompletionStats() {
+        return statsService.getCompletionStats(toDoList);
     }
-
-    private String formatMillisToStandardTime(long millis) {
-        long seconds = millis / 1000;
-        long minutes = seconds / 60;
-        long hours = minutes / 60;
-        long days = hours / 24;
-
-        seconds %= 60;
-        minutes %= 60;
-        hours %= 24;
-
-        return String.format("%02d:%02d:%02d:%02d", days, hours, minutes, seconds);
-    }
-
-
-
-
-
-
 }
